Implement recipe recording
Recipes are recorded and saved in a json file when the potion is saved. The json file can be visualized elsewhere. I've been using a notebook for a quick view, but I'll probably make a better tool for it later, or best case, make a way to display it in-game. Added support for rotation, health recording, and salts, and a way to see if the potion is currently teleporting. I ended up sort of fixing one of the bugs in the game where heating at a whirlpool doesn't add any notes to the recipe. I also made quicksave and quickload work properly with this data, though it will not work if you just load into a save file in the middle of a potion. Also added a visual display for health.